requirements.txt
torch==2.6.0
torchvision
torchaudio
transformers>=4.30
diffusers>=0.21
accelerate>=0.20
safetensors>=0.3.1
omegaconf>=2.1
einops>=0.6
Pillow>=9.0
numpy>=1.21
scipy>=1.7
imageio>=2.9.0
imageio-ffmpeg>=0.4.5
runpod>=1.0.0
cloudinary>=1.33.0
requests>=2.28.0
python-dotenv>=0.19.0
flash-attn>=2.0
